Add support for WPA Supplicant pass-through.
Implement a solution based on `socat(1)` for mirroring the guest's
`wpa_supplicant(8)` control socket interface on the host, because
it cannot be done through VirtFS/9p.  This requires a guest system
that can open the respective TCP ports per the shared configuration
file, `wpa_ctrl.conf`.

The `wpa_supplicant.conf(5)` configuration file is moved to a
dedicated directory which could be mounted writeable and therefore
updated from the guest when instructed through the control
interface.

Because the control socket interface has to be explicitly
configured for WPA Supplicant, disable this feature by default.