Make dependency on `socat` optional.
WPA Supplicant pass-through is not configured by default, so the
user should not be forced to install it as a dependency.  Make the
the code a bit more fail-safe by not trying to use `socat` if it
is not present on the host.  (The user may enable WPA pass-through
later without installing `socat`.)

While here, improve the related log messages.