Employ more resilience.
The `start`, `stop`, and `restart` operations should consider the
proper management of the networking interfaces and the control
socket forwarding besides the guest itself.  The goal should always
be to make sure that the application is in the desired state and
apply all the changes to reach that, i.e. create or stop components
as needed.