Automate management of the suspend/resume recovery method.
There are multiple ways to implement recovery on the suspend and
resume events, and utilize the related `rc(8)` facilities to
handle this in a more seamless manner.  Note that the `rc(8)`
framework supports `resume` only, so the `devd(5)` configuration
is kept around to bridge the gap on the side of `resume`.

Inspired by:	Maurizio Vairani <maurizio1018@gmail.com>