Add support for configuring multiple PCI pass-through devices.
Sometimes it might be necessary to work with multiple pass-through
devices.  For example, when the guest wants to talk to the
Bluetooth controller of the wireless device and this requires the
communication with the USB xHCI controller as well.

Allow the `passthru` configuration option in `bhyve.conf` to hold
multiple PCI device slot/bus/function values.  This extension is
backward compatible.

Sponsored by:	Universal Devices