Sunset bhyve+ support alongside FreeBSD 12.
It was a promising experiment but it did not seem to be important
enough to get continuous support and development.  Now, that
FreeBSD 12 has reached its end-of-life, there is no strict need
for this utility.  Remove the corresponding configuration
parameters for the sake of simplicity.

Note that each of the script variable that controls the location
of the respective bhyve component could still be overridden from
the environment.