Add syntax check for PCI pass-through configuration.
There was no check to ensure that the `passthru=` configuration
option in the `bhyve.conf` file is in the expected format, but
that was passed blindly through a sequence of commands.  Though
the format is documented in the file, it cannot be safely assumed
that it will be given properly, possibly by mistake.

At the same time, use this opportunity to provide a better error
message for the user and cause less frustration.

Fixes #82