Make documentation more explicit on certain details.
Previously, it was not emphasized that Wifibox is not a drop-in
replacement for the FreeBSD wireless networking stack, which
could have created confusion and false expectations.  Add
sentences to both the README and the manual page to clarify that
this is not the case.

Add more pointers to `wifibox-guest(5)` for the manual page
to help the users understand that the most part of the
configuration actually depends on how the guest is constructed.

Recommend the user to verify the sample configuration before use,
and place a hint on where to increase a logging verbosity for
diagnostics.

Fixes #83