Be more adaptive about the location of the `vmm` kernel module.
It is not always the case that the kernel modules are located
under the `/boot/kernel` directory.  For custom kernels, that is
something else (though usually somewhere within `/boot`) which
makes the script fail to function properly.

On initialization, query the currently lodaded kernel itself
about the path of its accompaining modules and use that
information for finding `vmm(4)`.

Fixes #86