Make the priority of the `bhyve` process configurable.
Currently, the guest is run with the same on priority level as any
other process.  This does not always work well for handling the
wireless traffic, since firewalling and other networking functions
in the kernel are run with higher priorities.

Employ `nice(1)` to adjust the priority of the managed `bhyve`
process that is responsible for running the virtualized guest, and
expose its prority setting as a configuration value in
`bhyve.conf`.  This value has its own scale, which independent
from what `nice(1)` has in order to make it possible to replace
this tool for others, e.g. `rtprio(1)` later on.

Suggested by:	Anton Saietskii <vsasjason@gmail.com>