Implement validation for values in `bhyve.conf`.
It is not ensured that elements of the `bhyve` configuration are
in either the expected format or range.  This is going to show up
sooner or later and stop the guest from working, but it is more
friendly if these mistakes are reported directly to the user.

This commit is inspired by 221b2c07.