Remove explicit management of `nmdm(4)` devices
Character devices under the governance of the `nmdm(4)` driver are
created on demand on first access therefore it does not make sense
to check for their presence and remove them based on the response.
Similarly, this behavior of null-modem devices makes it unncessary
to create them explicitly before `wifibox` gets to use them.

In addition to that, removing existing `nmdm(4)` devices ahead of
time externally causes the kernel module fail to unload.  This is
clearly a bug in the base system itself, but let us reduce the
chances and simplify the code of `wifibox` itself.

Remove both the `create_nmdm` and `destroy_nmdm` functions as well
as the option to restart the console, as an attached
functionality.

Fixes #104