Correctly terminate bhyve VM using kill SIGTERM.
Previously, a `kill -SIGTERM` would be requested of the daemon, which would
forward the signal to wifibox. wifibox would not forward the signal to bhyve,
which requires the signal to shutdown cleanly. Instead, wifibox now kills the
pgid of the daemon, which will kill all of the children of daemon as well.

Also call echo and tr with variables ${ECHO} and ${TR} respectively.

Also explicitly unlink the tap interface from the wifibox interface, and
consistently use subshell format, standardize indentation, and use kill -TERM
instead of kill -SIGTERM.