Adaptively wait on stopping the guest.
The static 3-second timeout for waiting the guest on stop may not
be enough for the guest to complete its shutdown sequence.  This
can be important for letting the guest to run certain tear-down
functions and prevent from leaving the hardware in an troublesome
state.

Instead, after the `TERM` signal is sent to initiate the graceful
termination of the guest, periodically check back on its state for
a while (configurable by the user, defaults to 10 seconds) and
either consider it stopped or make an attempt to forcefully stop
the guest to avoid the infinite wait.

This is also some form of mitigation of the bug that was uncovered
in 95312ad5.  Previously the forceful shutdown was always executed
because the ACPI power-off did not work.  But with that change,
it now often becomes unnecessary and it actually leads to a race
condition and error with bhyvectl(8).