Make inclusion of the `devd.conf(5)` extension optional.
FreeBSD versions starting from 14.0 can now properly trigger the
`suspend` action for `rc(8)` scripts [1] hence the workaround for
that is no longer needed.  Implement an automated detection of
this feature in the build process and lose the fix depending on
the OS version.

Note that since the original change in `src` did not bump the
`__FreeBSD_version`, it is not possible to tell exactly if the
userland has the necessary support.  Thus an approximate version
is used to err on the side of caution.  Fortunately, that is not
a problem if the suspension logic is run twice, because that is
idempotent.

Fixes #120

[1] https://github.com/freebsd/freebsd-src/commit/2cf8ef5910fd3754f8021f9c67fde6b9d9030f33