Send the `TERM` signal to the guest directly.
On stopping the guest, a `TERM` signal is sent for the whole
process group that was created for the VM manager.  This means
that the VM manager is also stopped which should be taking care of
destroying the corresponding VM.  In result, the VM is left behind
and not cleaned up eventually.

Switch over to sending the `TERM` signal to the guest itself.  The
VM manager will be notified of the guest stopping anyway, since
`bhyve` is going to translate `TERM` to an ACPI power-off event.
This would then gracefully let the manager to take care of the
associated clean-up tasks, including destroy the VM.  At the same
time, focus the guest itself to see if it is running, and do not
care about the manager.

This is an alternative approach to implement #118.