Avoid race conditions on restarting the guest.
The `ppt` driver cannot be cleared until the VM runs so move the
respective `devctl(8)` call back to the end of `destroy_vm`.
Restore the wait period after the VM is destroyed to make sure that
the device is released.

More importantly, do not let `vm_stop` to call `destroy_vm` but
leave it the VM manager to take care of the clean-up tasks.  The
VM manager should be unblocked to do that after the guest either
stops by itself (due to the ACPI power-off event) or forcefully
stopped.  In addition to that, wait for the VM manager to stop too,
because it should not be launched again (on a different thread)
until all the clean-up tasks have been completed.