Extend CI build checks to cover manual pages.
Configure an additional CI job to run some basic formatting and
style checks on the changes to the manual page.  As part of that,
address all the identified issues there so we could start with a
clean slate.

At the same time, make running ShellCheck dependent on changes to
the script itself, so it will not be run unless that is necessary.