Consider built-in kernel modules on checking their presence.
The `assert_kmod_loaded` function expects kernel modules to be
loaded from a specified location and it fails to work if the
module in question has been actually added to the kernel.  This
case was not considered before because both `nmdm` and `vmm` are
shipped as optional, loadable modules with the stock FreeBSD
kernel, and that is usually not an issue.

Relax this restriction and allow working with built-in kernel
modules.  This way users with customized kernels would be
supported.  The original motivation behind this check is to ensure
that the loadable modules are created for a compatible kernel.
But that implicitly holds when they are included there.

Fixes #136